+++ laravel/app/Http/Controllers/StockController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Order;
-use App\Inventorie;
+use App\Inventory;
 use Carbon\Carbon;
 
 class StockController extends Controller
@@ -16,13 +16,11 @@ class StockController extends Controller
 
     public function order()
     {
-        $now = Carbon::now();
-        return view('order', compact('now'));
+        return view('order');
     }
 
     public function order_index(Request $request)
     {
-        $now = Carbon::now();
         $item_id = $request->input('item_id');
         $delivery_user_id = $request->input('delivery_user_id');
         $order_start = $request->input('order_start');
@@ -30,7 +28,7 @@ class StockController extends Controller
 
         $order_indexes = Order::orderIndex($item_id,$delivery_user_id,$order_start,$order_end)->get();
 
-        return view('order', compact('order_indexes','now','item_id','delivery_user_id','order_start','order_end'));
+        return view('order', compact('order_indexes', 'item_id', 'delivery_user_id', 'order_start', 'order_end'));
             
     }
 
@@ -46,19 +44,17 @@ class StockController extends Controller
 
     public function stock()
     {
-        $now = Carbon::now();
-        return view('stock', compact('now'));
+        return view('stock');
     }
 
     public function stock_index(Request $request)
     {
-        $now = Carbon::now();
         $item_id = $request->input('item_id');
         $delivery_user_id = $request->input('delivery_user_id');
         $status = $request->input('status');
 
-        $stock_indexes = Inventorie::stockIndex($item_id,$delivery_user_id,$status)->get();
+        $stock_indexes = Inventory::stockIndex($item_id,$delivery_user_id,$status)->get();
 
-        return view('stock', compact('stock_indexes','now','item_id','delivery_user_id','status'));
+        return view('stock', compact('stock_indexes', 'item_id', 'delivery_user_id', 'status'));
     }
 }
