+++ laravel/resources/views/common.blade.php
@@ -4,16 +4,16 @@
     <title>{{ config('app.name') }}</title>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" type="text/css" href="{{ asset('css/ads.css') }}">    
+    <link rel="stylesheet" type="text/css" href="{{ asset('css/ads.css') }}">
   </head>
 
   <body>
     <header>
-      @yield('header')     
+      @yield('header')
     </header>
       @yield('content')
     <footer>
-      @yield('footer') 
+      @yield('footer')
     </footer>
   </body>
 </html>
